+++ src/server/adapter/parser/dql_examples/dynamicres.dql
