+++ src/server/adapter/parser/tokenizer/tokenizer.go
@@ -1,7 +1,7 @@
 package tokenizer
 
 import (
-	tok "github.com/domain-query-language/dql-server/src/server/adapter/token"
+	tok "github.com/domain-query-language/dql-server/src/server/adapter/parser/token"
 )
 
 type Tokenizer interface {
